FastScripts module
Declaratively setup user scripts for FastScripts, in order to run the [script] that has Safari prompt prior to quitting. The module reuses some logic from the [launchd module][launchd] which depends on [write-text][write-text]. Hopefully this is not a house of cards that will require later rewriting. It does obviously couple the NixOS configuration to internal modules, but those modules are eventually exposed as options in public-facing ones.

[script]: https://daringfireball.net/2020/01/quit_confirmation_for_safari_on_macos
[launchd]: https://github.com/nix-darwin/nix-darwin/blob/44a7d0e687a87b73facfe94fba78d323a6686a90/modules/system/launchd.nix
[write-text]: https://github.com/nix-darwin/nix-darwin/blob/master/modules/lib/write-text.nix